fix: multiple bug fixes and UI improvements
- Fix publish recipe dialog loading hang (finally block for _isPublishing)
- Style rating count badge similar to rating badge (icon + blur background)
- Make card title max 1 line with text truncation
- Replace visibility toggle with collection management in FoodDetails
- Change rating from int to decimal slider (0.0-10.0 with 0.5 steps)
- Fix global recipe missing metadata on publish (prep time, cook time, etc.)
- Fix 'Kopier og tilpass' to create independent copy without HelloFresh data
- Remove HelloFresh recipe creation toggle from AddFood page
- Remove empty image container when browsing HelloFresh recipes
- Fix 'Legg til i samling' dialog on friends page (MudBlazor 8.15 compatible)

+++ StorhaugenEats.API/Models/ActivityFeedItem.cs
@@ -58,7 +58,7 @@ public class ActivityFeedItem
     public User User { get; set; } = null!;
 
     // Static factory methods for creating activities
-    public static ActivityFeedItem CreateRatingActivity(Guid userId, Guid recipeId, string recipeName, int rating, string? imageUrl = null)
+    public static ActivityFeedItem CreateRatingActivity(Guid userId, Guid recipeId, string recipeName, decimal rating, string? imageUrl = null)
     {
         var metadata = new Dictionary<string, object?>
         {
