+++ StorhaugenEats.API/Services/IRatingService.cs
@@ -7,7 +7,7 @@ public interface IRatingService
 {
     Task<Rating?> GetUserRatingForRecipeAsync(Guid userId, Guid globalRecipeId);
     Task<IEnumerable<Rating>> GetRatingsForRecipeAsync(Guid globalRecipeId);
-    Task<Rating> UpsertRatingAsync(Guid userId, Guid globalRecipeId, int score, string? comment = null);
+    Task<Rating> UpsertRatingAsync(Guid userId, Guid globalRecipeId, decimal score, string? comment = null);
     Task<bool> DeleteRatingAsync(Guid userId, Guid globalRecipeId);
     Task<List<UserRatingDto>> GetGlobalRecipeRatingsForUserAsync(Guid userId, int skip = 0, int take = 50);
